refactor(dashboard): move generic coercion helpers to lib/json-coerce.ts
Move the pure unknown→typed coercion helpers from components/common/normalize.ts
into lib/json-coerce.ts and re-export them from normalize.ts. This removes the
layer-cycle blocker that forced lib/fusion-meta.ts to reimplement asString/asRecord.

- lib/json-coerce.ts: new SSOT for asString/asNumber/asBoolean/asInt/
  asStringArray/asRecordArray/asNullableString/asRecord/asStringList/extractArray.
- components/common/normalize.ts: re-exports the moved helpers; existing component
  consumers keep their import paths.
- lib/fusion-meta.ts: imports asString/asRecord from lib/json-coerce.ts and drops
  its local duplicates. Keeps a private asNumber that also accepts numeric strings
  because fusion wire data emits token counts both ways.

pnpm typecheck and the touched test suites pass (240 tests).
